requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.9.0
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
